+++ packages/local-workspace/index.ts
@@ -13,7 +13,6 @@ export {
   STATES_DIR_NAME,
   locateWorkspaceRoot,
   createEnvironmentSource,
-  getAdaptersConfigTypesMap,
 } from './src/workspace'
 export {
   workspaceConfigSource as localWorkspaceConfigSource,
